backend/requirements.txt
Django==5.0.6
requests==2.32.3
djangorestframework==3.15.1